Added random background colors and favicon
I like the favicon and the new background colors, but definitely NOT
the code for the new background colors. Like, jeez, I could easily make
a double nested loop with 2 arrays (one containing
“-moz-“,”-o-“,”-webkit-“,”” and the other the id names) and make the
long ass code a lot shorter.  And even the variable values could easily
be a loop and array.

This WILL change soon